Add rudimentary example and build system (#1)
Example apps are located in their own directory to make it easier to
tell them apart from libraries and binaries that can be reused as is.

The first example is `hello_world` because it is the simplest app that
demonstrates that this method of building apps works. It does
essentially the same thing as its namesake from the C example, so it
might as well share its name to facilitate comparison.

The app is located in the `apps/` directory to make it easy to separate
examples from reusable libraries and binaries, which will be located in
a separate directory.

The manifest.json file is adapted from the corresponding C example
with some modifications:
* `embeddedSdkVersion` is omitted to keep boilerplate to a minimum.
* `schemaVersion` is set to a version that maximizes compatibility
  while allowing `user` to be omitted.
* add a newline at the end to ensure all tools recognize the last line.

The version in `Cargo.toml` is set to match `version` in
`manifest.json` because why not.

The LICENSE is adapted from the hello-world example in the
acap-native-sdk-examples repo.

`cross` is used because it spares us decisions about how to build and
tag docker images. Alternatives include
* `cargo-acap` rejected because it does not support recent manifest
  features and to link against the SDK we would need to manage docker
  images anyway.
* plain `cargo` because we would need to manage docker images.

However, I expect we will move away from using `cross` as we have more
time to think about these questions.

Run CI both on pull-request and merge because currently two pull
requests can be independently verified and merged or rebased without
running the tests again. If they are incompatible, then a third pull
request based on top of both could not easily distinguish between a
problem it has introduced and a problem that already existed on main.