Add rudimentary example and build system and helpful shortcuts (#1)
The first example is `hello_world` because it is the simplest app that
demonstrates that this method of building apps works. It does
essentially the same thing as its namesake from the C example, so it
might as well share its name to facilitate comparison.

The app is located in the `apps/` directory to make it easy to separate
examples from reusable libraries and binaries, which will be located in
a separate directory.

The manifest.json file is adapted from the corresponding C example
with some modifications:
* `embeddedSdkVersion` is omitted to keep boilerplate to a minimum.
* `schemaVersion` is set to a version that maximizes compatibility
  while allowing `user` to be omitted.
* add a newline at the end to ensure all tools recognize the last line.

The version in `Cargo.toml` is set to match `version` in
`manifest.json` because why not.

`cross` is used because it spares us decisions about how to build and
tag docker images. Alternatives include
* `cargo-acap` rejected because it does not support recent manifest
  features and to link against the SDK we would need to manage docker
  images anyway.
* plain `cargo` because we would need to manage docker images.

However, I expect we will move away from using `cross` as we have more
time to think about these questions.

Run CI both on pull-request and merge because currently two pull
requests can be independently verified and merged or rebased without
running the tests again. If they are incompatible, then a third pull
request based on top of both could not easily distinguish between a
problem it has introduced and a problem that already existed on main.
Initial commit

Expose eap-install.sh commands as make shortcuts to document how they
can be invoked and to make it easier to run them. The names of
parameters are meant to be aligned with existing acap examples and are
taken from the web-server examples.

Also add a `run` shortcut to document how to run a program in a way
similar to how it would run when the app is started and to make it
easier to do.

Unlike previous shortcuts that get their name from the corresponding
command in `eap-install.sh` the `run` shortcut is meant to evoke
 `cargo run` and this is why it builds the executable as well.

Forward `RUST_LOG_STYLE` to remote process to allow users to enable
color by setting it to `always.

Git LFS is not used for `Cargo.lock` because so far it is only 400 lines
and not everyone has Git LFS installed. `package-lock.json` files can
grow quite large, 36k lines for one app I know, and if our lock file
approaches that we may want to reevaluate Git LFS.