Add licensekey wrapper and example
The example is named the same as the corresponding C-example to
highlight that they do essentially the same thing.

Benefits of the ergonomic API over the flexible API is mainly that
* one redundant argument has been removed, and
* the return type is easier to user.

The two are nonetheless separated to establish a pattern that I am
fairly confident will be useful e.g. for axevent.

The function will also be useful if we add a `verify!("manifest.json")`
procedural macro.

Bindings.rs are ignored to signal that this is a hack and because they
can trivially be regenerated.

Exclude crates that depend on the ACAP APIs when targeting host because
they will fail.

Another option is using conditional compilation to make sure they build
on all targets but this ends up being hard to read and I have yet to
find a way to improve it. Notable pain points include
* links in inner docstrings break and the only solution I have found
  for making them conditional requires putting parts of the docstring
  inside a #[] which makes it hard to read and I did not even try it.
* Readability is important in the sample apps and bloating them with
  conditional configs to make them build for host does not make sense.

Run checks for aarch64 as well because it is nice to keep things
working on host if there is no good reason not to. It would be nice if
eventually all crates had some unit test suite that runs on host for
fast feedback and robust CI.

Don't run checks for armv7hf because I expect the results will be
highly correlated with aarch64 and it would increase CI time a lot.

Tests are not run for non-host targets because even though cross
supports emulation the necessary shared libraries would not be
available.

Change-Id: I2f0572d4aa14904d86b9b181528225b06029fb76