Add helpful make shortcuts (#3)
Expose eap-install.sh commands as make shortcuts to document how they
can be invoked and to make it easier to run them. The names of
parameters are meant to be aligned with existing acap examples and are
taken from the web-server examples.

Remove some output from the commands since attempting to follow these
instructions most likely would not work.

Prefix commands with `@` primarily to avoid printing the misleading
instructions that we are trying to remove.

Also add a `run` shortcut to document how to run a program in a way
similar to how it would run when the app is started and to make it
easier to do.

Unlike previous shortcuts that get their name from the corresponding
command in `eap-install.sh` the `run` shortcut is meant to evoke
 `cargo run` and this is why it builds the executable as well.

Forward RUST_LOG_STYLE to remote process to allow users to enable color
by setting it to `always.