feat: Add metadata broker wrapper and example
The example does roughly the same as
consume-analytics-scene-description and it seems reasonable that we may
want and be able to align them. But that name is too long so we use the
shorter name from the manifest consistently instead.

Since we cannot have multiple crates with the same name we might as
well align the name of the directory with the name of the app.

The rust version is updated to enable use of c-string literals to make
the code more readable.

Abort process on mdb error

We don't want the process to continue running in a bad state. Aborting
is what the C example does and in this sample app it may actually make
sense since without the mdb subscription the app does nothing. Any app
based on this example probably needs more sophisticated error handling
though.

Update ACAP Native SDK to 1.14 because at least 1.13 is needed for the
broker and we may as well upgrade all the way.

Relax closures to `FnMut`

This is necessary if we want to be able to drop the sender end, which
seems like a sensible thing to do if the receiver cannot keep up.
Alternatives considered are
1. Using blocking send; this seems like a bad idea since it will likely
   block the broker.
2. Logging an error; this seems like a bad idea since the receiver will
   not know that messages have been dropped.
3. Panicking; this is equivalent to logging an error in the case of
   unwind and leads to applications that degrade poorly in the case
   of aborting panics.
4. Explicitly relaying information about dropped messages e.g. using a
   wrapper type or sequence numbers; I don't want to push this
   complexity onto users and I suspect it would not be possible to
   implement without relaxing the closures to `FnMut` anyway.

Use manifest schema version 1.6.0 because metadata broker is
introduced in 11.9, so using a lower version does not provide any
benefits.