Track generated bindings (#4)
This has two effects, both of which serve to make it easier to write
and test documentation:
* it enables `cargo package` which is needed to test with docs.rs since
  we are using workspace.
* it enables `cargo doc` to build all crates for the host target. This
  may be slightly misleading because the crates can only be compiled
  for Axis devices, but the convenience of having some docs outweighs
  this in my opinion.

For the licensekey header the bindings are the same regardless of build
target and profile. For other headers this may not be the case, and in
that case we may want to be more specific about which build we copy the
bindings from in order for the noun targets to be reproducible.

Replace `copy_bindings` target with `check_generated_files` to ensure
bindings are kept up to date.

Add the `--always-make` option in CI to ensure that generated files
are remade even though they already exist. If they are not, then the
git check that follows would be futile.