Add support for bundling files
Because removing files in the build tree that have been removed in the
source tree is a big hassle, a single target is used to stage the files
that acap-build will bundle.

Because the make target depdends on otherfiles and updating
otherfiles/html/index.html does not invalidate the make target, we hack
the binary target to ensure that otherfiles are always copied.
This could be solved properly, but I think it would introduce more
complexity than it is worth at this point.