Demonstrate containerized builds
This should make it easier to build the example apps in CI systems
where setting up docker-in-docker is difficult. As desirable side
effect of this is that it becomes easier to build the example apps on
any system, including those where following the advanced setup may be
difficult.

To be able to build both inside and outside docker containers branching
is introduced in the makefile. The added complexity is unfortunate but
worthwhile in order to better support containerized builds. The goal is
to eventually replace most of the makefile with a cargo plugin.