Add acap-ssh-utils bin and lib crate
Benefits of having this include
* documents how to quickly deploy and test incremental changes.
* makes it easier for other tools to do so by taking care of the many
  details while presenting a friendly interface.

This will come in handy for implementing analogs of `cargo run` and
`cargo test`, both of which will cross-compile a binary and execute it
on target.

Exclude the new crate when checking other targets because supporting
other targets is not a priority, especially not for this crate that
will depend on `sshpass`, `ssh`, and `scp` for the foreseeable future
which makes it less portable.

Change the behavior of `make run` to always set the `RUST_LOG*`
environment variables to decouple the log level of the tool that runs
on host from the log level of the app that is running on target.

Always require a password because it is not possible to set up public
key authentication for users other than root, so we have to support
password authentication and supporting only one method makes both the
implementation and the user experience more straightforward.

Print information about environment from `hello_world` example because
that way it becomes a useful tool and it doesn't make the app
noticeably more difficult to understand.