chore: Factor out app_logging modules to lib crate
This change is prompted by wanting to demonstrate how to use library
crates from this repo in acap-rs-app-template.

But it also makes sense in the context of only this repo where verbatim
copies of the `app_logging` module are included in all current and
upcoming example apps; factoring it out will prevent the copies from
accidentally diverging.