Rename environment variables controlling the Makefile
Prefix also `ARCH` and `PASS` with `DEVICE_` to communicate that they
refer to the architecture and password of the device under test.

Prefix all the variables with `AXIS_` to make it less likely that they
will collide with variables used differently way by other programs.
This specific prefix is chosen to align with acap-native-sdk-examples
where `AXIS_DEVICE_IP` was recently adopted as the only placeholder
name for the IP address or domain name of a camera.