chore: Explain more make targets
The new "docstrings" don't add much information, but they make it
easier comparing the `Makefile` in this project to that in
acap-rs-app-template because the diff is smaller and they only differ
where there is a good reason for doing so.

Also add and remove some newlines; double newlines only precede
sections to make them easier to spot.