s: Serve static files on host
Serving static files on host shortens the feedback cycle since we don't
need to deploy anything to a device.

The way that static files are located will probably not work in general
once support for dynamically generated files is added to the build
system, but this example only uses static files, so it should not block
that feature.