fix(embedded_web_page): Put web page in `html/` (#30)
The `acap-build` tool establishes the convention that the `html/`
and `lib/` directories in the current working directory are special and
always included. This seems convenient, so going forward the tools and
examples in this project will follow that convention.

Disable unit tests on host because it is not obvious how to implement
`html_dir`.