s: Rewrite web page
Because I don't know of an official way for the web page to figure out
what access policy it is authorized for I add an input for selecting it.
Attempting to use a more restricted policy than the user is currently
authorized for will trigger an authentication challenge.

For simplicity the web page always uses the websocket as "anonymous"
because this should always work. Compared to using the selected policy
this is simpler because the websocket does not have to be closed.