Make the `build` verb build for only one target
I rarely find myself wanting to build for targets other than the device
that I am currently testing with, and having to wait twice as long as
necessary is annoying.

I expect that CI workflows will still want to build for multiple
targets, but that it should be a relatively small effort to call make
twice since the developer only has to write that code once.

A side effect of this change is that CI now only builds apps other than
`hello_world` for one target. This could easily be fixed by adding a
couple of lines in `CI.yml` but I think this is probably desirable
since I expect the compilation success to be highly correlated.