Adopt dev container as primary environment
Going forward the dev container is the primary supported dev
environment, so it is important to ensure that it works.
The "containerless" environment should also work but is tested less
rigorously in CI because I expect the results of the omitted tests to
be highly correlated with the results from the dev container and I
don't want to make CI slower than needed. Furthermore this is my
default environment, making it more likely that I would discover
problems during development.

All container / no-container branching is consolidated into one place
mostly because it will make later changes easier since they too will
us the `DOCKER_RUN` macro in one of the branches.

In `Dockerfile`:
- install libglib2.0-dev because without it `check_docs` fails.
- install `iputils-ping` because without it the `eap-install.sh` script
  fails because it thinks the device cannot be reached.