s: Use name "default" to get credentials
Use the name "default" when getting credentials since the exact value
does not matter but it cannot be empty.
In practice this name is only used as a suffix for the returned
username. The prefix is `a{uid}-` where `uid` is the user id of the
user the app runs as. Since the app should (must?) run as a dynamic
user to access this API there is no risk of colliding with another
app, so as long as this app does not use multiple credentials
concurrently the exact value is not important.