Make the `build` verb build for only one target (#31)
I rarely find myself wanting to build for targets other than the device
that I am currently testing with, and having to wait twice as long as
necessary is annoying.

I expect that CI workflows will still want to build for multiple
targets, but that it should be a relatively small effort to call make
twice since the developer only has to write that code once.

A side effect of this change is that CI now only builds apps other than
`hello_world` for one target. This could easily be fixed by adding a
couple of lines in `CI.yml` but I think this is probably desirable
since I expect the compilation success to be highly correlated.

I considered keeping the old behavior but with a new target named `all`
because I received the feedback that it was surprising that:
- the `all` target does not exist.
- the `all` target is not the default.

I'm choosing to not do that for now because what "all" means in the
context of this project is a little bit unclear. Questions include:
- should it build for all architectures
- should it build the docs, which also is an artifact that the project
  provides.