s: Focus interface on `.eap` files.
The crate is focused around the package as a unit that is patched and
run. Since every ACAP developer have systems set up to produce `.eap`
files I expect this to be an accessible interface.

I previously considered an interface where files individual files could
be replaced. That was driven by
- having worked with an app for which incremental compilation worked
  poorly and for fast iterations it was necessary to build and patch
  only part of the app. I know think this would be better solved by
  improving that build system, and it is not something that this tool
  should attempt to accomodate.
- having a build system for apps written in rust that does not produce
  `.eap` files when compiling tests. I have since rewritten the build
  tool so be more `.eap` centric and I think this is the easier
  interface to work with when combining tools.