Improve acap-logging docs
The crate level documentation is taken from `README.md` to make it
easier to ensure that crates.io, docs.rs and GitHub all show sensible
documentation with minimal effort.

Also update the `hello_world` example since it would not display any
output by default when run in a terminal. To increase the utility of
the app a bit further without increasing the complexity it is used to
also demonstrate the effect of logging at other levels and of printing.