Improve acap-logging docs (#41)
The crate level documentation is taken from `README.md` to make it
easier to ensure that crates.io, docs.rs and GitHub all show sensible
documentation with minimal effort.

The documentation test is annotated with the `no_run` attribute because
it fails to create a unix socket logger in GitHub Actions. In GitHub
this causes it to not be syntax highlighted but I'm hoping crates.io
will pick it up anyway.

Also update the `hello_world` example since it would not display any
output by default when run in a terminal. To increase the utility of
the app a bit further without increasing the complexity it is used to
also demonstrate the effect of logging at other levels and of printing.

Also add a table of library crates provided by this project and where
to find their documentation. This makes it easier to get an overview of
what is available and where.