Remove partially containerized workflow (#42)
The motivation for this change is to reduce the test burden early on in
development. I previously thought that hiding the containerization was a
feature, which motivated the complexity, but the feedback I have
received so far is that there should be no containerization and that
everything should run in a development container.

In `install-rust.sh`:
- Move the file to keep the number of directories in the repository root
down.
- Rename the file with a hyphen to be consistent with many other files,
including:
    - `environment-setup.sh`
    - `rust-toolchain.toml`

In `install-venv.sh`:
- Depend on a global installation of python because:
- Debian 12 and Ubuntu 22.04 both allow a compatible version of python
to be installed with `apt-get`.
- `pyenv` can be used on most systems to install exactly the python
version that is expected.
    - Installing Python locally uses additional storage and time.
- Depend on a global installation of Rust because:
    - Rust can seamlessly manage multiple, concurrent toolchains.
    - Installing Rust locally uses additional storage and time.

In `rust-toolchain.toml`:
- Add list of targets containing supported camera architectures to make
it easier installing Rust, components, and targets using the `rustup
show` command.

In `devcontainer.json`:
- Remove the `CARGO_HOME` directory because it is not set in the
container image, so the default value will be used. This is equivalent
to what was previously set explicitly.