Add acap-ssh-utils bin and lib crate (#19)
This is primarily intended as a building block for a more tightly
integrated development tool, where it will come in handy for
implementing analogs of `cargo run` and `cargo test`, both of which
will cross-compile a binary and execute it on target. But, since it
solves a common problem among ACAP developers using any language, it is
factored out into its own lib crate with a thin binary crate wrapper.

The crate is focused around the package as a unit that is patched and
run. Since every ACAP developer have systems set up to produce `.eap`
files I expect this to be an accessible interface. An earlier version
allowed individual files to be replaced. This was driven by working
with build systems, including my own, that did not package tests as
`.eap` files or were slow to produce `.eap` files. I now think that
these issues are better addressed in the build system since the action
of creating the tar archive should in theory be cheap. 

In `install-system-packages.sh`:
- Add `sshpass` because it is needed by `acap-ssh-utils` which in turn
  needs it because I have not found a way to add authorized keys for
  users other than root.

In `.gitignore`:
- Add `/.env` because defining the `AXIS_DEVICE_*` variables in a file
  makes it reduces the friction of rebuilding the dev container and to
  run tests.

In `Makefile`:
- Export the `AXIS_*` variables to make sure they are available for
  scripts when the default values are used. This is particularly useful
  for the password argument since it allows programs such as
  `acap-ssh-utils` to pick this up from the environment, which is
  better from a security perspective (that tool still passes it as an
  argument to other commands though).
- Always set `RUST_LOG` and `RUST_LOG_STYLE` when running on target
  because I found that most of the time this is what I wanted and
  oftentimes I did not want to set it for `cargo-acap-build` and
  `acap-ssh-utils` at the same time.