feat: Add `device-manager` bin and lib crate
Being able to quickly restore a device to a known baseline makes it
easier for me to do rigorous testing; removing apps manually is a
hassle and may not be enough to undo the effects that development has
had on a device either because I changed some setting or because an app
did.

The variant of VAPIX used has diverged from that which was integrated
as its own crate. Notable differences include:
- `acap-vapix` uses builder style public interfaces. These are closer
  to what I expect the final result will look like since they allow
  greater flexibility in use and evolution.
- `acap-vapix` draws the line between request and client in a way that
  gives the APIs less control over things like headers. This is better
  aligned with how many APIs since they are ignorant of headers,
  including authentication headers, which is something a user likely
  wants to control independently of which API is being used.
- `device-manager` is no longer parsing AJR responses as enum because
  this resulted in more boilerplate and is not helpful when the
  transport is already request-response based as HTTP is. The reason
  for using enums was that one intent of protocols like AJR is that
  they should be transport agnostic and work over WebSocket just as
  well as HTTP. As of writing I'm leaning towards a third framework
  that separates the envelope parsing and serialization from that of
  the data, errors, and parameters.

Because I am happy with neither branch and I don't want to refactor
bindings twice, they are not moved except for parameter management
which is less dependent on diverging code, has an API that is easier
to adapt, and is more in demand.