feat(axevent)!: Improve callbacks
The summary of this commit is poor because I am being lazy and changing
two things.

Box is removed from the type of subscription callbacks to make the
interface nicer; the user does not have to type out `Box::new(...)`
and there will generally be less indentation. One potential downside
of this is that monomorphization of generic functions tend to increase
the binary size and I suspect that callbacks are particularly bad since
each callback gets its own type.

The type of declaration callbacks is constrained to make avoiding
leaking the callbacks easier. It is probably possible to convert the
`FnOnce` to an `FnMut` the same way we have done in other places but
because the argument is a `Box<dyn ...>` this is not as straightforward
and since I'd like to remove the box from the interface entirely it is
not a detour I want to take.

The `CallOnDrop` type is introduced to make it easier to ensure that
callbacks are freed exactly once. Though convenient I think this
pattern seems dangerous since it is not immediately obvious where the
unsafe code is being called.