f: Consistently list `LICENSE` as prerequisite
Essentially any target that uses `cargo-acap-build` needs this
prerequisite because it will try to use `acap-build` which fails if
there is no `LICENSE` file.

The cause of it being omitted is probably that I didn't think enough
about which targets needed it and since `make build` usually precedes
`make run` and `make test` without cleaning the repository between,
these targets appeared to work.