Remove extra comma in devcontainer.json (#69)
Because with it the document is not valid JSON. Technically the
document does not have to be valid JSON since dev containers use JSON
with Comments (jsonc). However, we don't use any comments so we might
as well make it valid json to make IDEs and other tools that expect
`.json` files to be valid JSON happy.