chore: release cargo-acap-sdk 0.1.0
I do this not because the state of the tool has crossed some threshold,
but because it is being used and having versions makes it easier for
users to install and it makes it easier to talk about which version is
being used or when things change.

I do not plan to publish this version because I am still unsure about
the name.