feat(acap-vapix)!: return custom error from basic device info
One immediate benefit of this is making the authentication detection
more robust (in theory; I have not actually had any problems with it
so far). The real reason is to establish a pattern for propagating
errors in general and auth errors in particular when adding more APIs.

Testing has primarily consisted of running some of the development
tools:
- `device-manager reinit`
- `cargo-acap-sdk remove`

`crates/acap-vapix/src/http.rs`:
- Add a `downcast` method speculatively to allow users an escape hatch
  should they want to try to classify the error themselves. If it can
  no longer be implemented in the future then returning an error always
  is probably not a breaking change, even if it may break things.
- Add both authentication and authorization kinds to make it clear to
  users that they are different. The distinction is often ignored and
  HTTP does not make it easier by returning UNAUTHORIZED when the user
  needs to authenticate.