chore: use Rust implementation of mkhelp
Intended benefits include:
- Faster (`make help` goes from 80ms to 3ms)
- One step closer to having no Python dependencies (While it would be
  nice to remove the virtual environment I still think Python can have
  a place in Rust ACAP app development, especially for tests and misc.
  scripts).

One downside is that the Rust implementation is not compatible with
my `kindly` program that provides enhanced tab completions for
`Makefile`s, but I wasn't going to push that onto `acap-rs` users
anyway.

I admit that the utility of this tool is pretty low, and it could
probably be removed, but it gives me joy so as long as it does not get
in the way for someone else I will keep it.