Add `Cargo.lock` to generated files
Without this, updating `Cargo.lock` in a way that is incompatible with
the manifest files would not cause CI to fail, despite the use of the
`--locked` option.

My theory is that when `cargo-acap-build` is invoked, directly from the
`Makefile` or indirectly via `cargo-acap-sdk`, it uses `cargo metadata`
which updates the lockfile, causing later `cargo` invocations to
succeed even when using the `--locked` option.

Disproved theories include:
- The `cargo install` commands in `install-venv.sh` update the lockfile
  (checking the diff immediately after this step reveals no diff)
- Some other action, prior to `make check_all` update the lockfile
  (checking the diff immediately before this step reveals no diff)

`Makefile`:
- Use `cargo metadata` because it reconciles the lockfile without
  updating packages as `cargo update` would, and faster and with fewer
  side effects than something like `cargo build`.
- Force the target to make sure it the lockfile is updated and the
  check works as intended, even if run before or without the other
  checks.