feat(axevent)!: Use generic callbacks and remove leak (#59)
Box is removed from the type of subscription callbacks to make the
interface nicer; the user does not have to type out `Box::new(...)`
and there will generally be less indentation. One potential downside
of this is that monomorphization of generic functions tend to increase
the binary size and I suspect that callbacks are particularly bad since
each callback gets its own type. Below are measurements on the example
app, but I don't think they are useful because the example app only
uses each function that takes a callback once:

- Release profile: 5520K -> 5524K
- App profile: 1020K -> 1020K

Other possible side effects, that I have not investigated, include:
- Better, worse or different compiler errors.
- Harder to package as rlib due to more generic methods.

Declaration complete callbacks are no longer leaked. This is arguably
two steps forward and one step back, because to accomplish this the
type of the callback was changed to `FnMut`. However, reversing this
interface change should not be breaking and I still haven't confirmed
that treating the callback as `FnOnce` is even safe.

The `Deferred` type is introduced to make it easier to ensure that
callbacks are freed exactly once. Though convenient I think this
pattern seems dangerous since it is not immediately obvious where the
unsafe code is being called.


`apps/send_event/src/main.rs`:
- Make the declaration callback `FnMut` even though I expect it to be
  called at most once because I have not confirmed this suspicion and
  implementing dropping it as `FnOnce` turned out to be difficult.
- Add a test ensuring that it continues to be possible to provide no
  callback. The main benefit of this is failing to compile if there is
  a regression, but running it makes it show up in test summaries which
  is nice.

`crates/axevent/src/flex.rs`:
- Guard the handlers with `Mutex` to keep the handler `Send` and `Sync`
  like the underlying C struct.