fix!: Make builds reproducible in dev-container (#99)
The main benefit of this is as a test case for the build tools,
especially if I go ahead and port `acap-build` to Rust; it will help
ensure that:

1. It is to accomplish reproducible builds.
2. Changes are not introduced unintentionally.

Keeping the output exactly the same would only be done to the extent
that it is helpful for the development process.

Ideally builds would be reproducible also outside of the dev-container,
but the binary differs. One reason is that paths to source files are
included and these include the cargo home directory which is typically
located in the users home directory. This can be addressed with the
`--remap-path-prefix` option to `rustc` which I successfully configured
globally on my machine like below (substituting <HOME> with my actual
home directory):

```toml
[build]
rustflags=['--remap-path-prefix=<HOME>.cargo=/usr/local/cargo']
```

Because I found no convenient way to configure it to affect only this
project, and because this configuration alone was not enough to get
reproducible binaries, I choose to not investigate further. When the
`trim-paths`[^3] feature drop hopefull this will be easier.

`Makefile`:
- Set `SOURCE_DATE_EPOCH` by default to make results reproducible
  across machines by default to reduce the number of attempts
  contributors have to make to get through CI.
- Set `SOURCE_DATE_EPOCH=0` to make it possible to detect when either
  the app or the build tools involved have changed and to reduce how
  frequently the file must be updated.
  This is suboptimal because it does not allow other tools to assume
  that all files are modified before this time and builds happen after
  this time, as the standard suggests[^2].
- Replace `aarc64` with `$(AXIS_DEVICE_ARCH)` to make it easier to
  explore a larger set of tests and, perhaps, eventually parametrize
  the tests in CI.
- Sort files before passing them to `sha256sum` to get consistent
  results as recommended by Reproducible Builds[^1].

`crates/cargo-acap-build/src/acap.rs`:
- Replace standard copy function to avoid preserving permission bits;
  Since files like `manifest.json` are tracked in git and the code is
  typically checked out on host, they will have their permission bits
  set on host causing even containerized builds to differ across
  machines.

[^1]: https://reproducible-builds.org/docs/stable-inputs/
[^2]: https://reproducible-builds.org/specs/source-date-epoch/#idm55
[^3]: https://doc.rust-lang.org/cargo/reference/unstable.html#profile-trim-paths-option