chore: Don't use clearly defined for license info (#101)
I recently struggled with merging a change because the checksums seemed
to keep changing. One potential explanation is that the LICENSE file
generation is not reproducible; when I test locally only 2/10 runs
produce identical license files for all apps. After this change 10/10
runs produce identical license files for all apps.

The problems may coincide with warnings about 429 and 502 responses
from clearly defined, but I have not confirmed if there is actually a
correlation.

There are a couple of issues in `cargo-about` vaguely related to the
problems I see, but they neither directly address the concern of
reproducibility nor seem to be worked by anyone else at the moment:
- 218
- 246

I compared the generated license for the `hello_world` app and the
differences do not seem catastrophic:
- Some licences have been reordered. These all seem to have had an
  instantiation of the license template replaced with the template
  itself.
- Notices about `aho-corasick` and `memchr` being dual licensed have
  been removed; The MIT license of both remain.
- A comment about `winapi` being dual licensed have been removed.

`Makefile`:
- Set `--fail` for good measure; It didn't cause any failures for me
  either with or without `no-clearly-defined` meaning I could not
  observe any improvements. But in theory it should be easier to detect
  that it was unnecessarily set than the other way around.

`about.toml`:
- Set `no-clearly-defined` because this stops `cargo-about` from
  consulting clearly defined, which seems to make the builds
  reproducible. Other potential benefits from this include lower risk
  of supply chain attacks and faster builds.