(chore): Add brief documentation of expected workflow
By omitting the decision element the layout gets simpler and easier to follow.
By also omitting the failed link the layout gets simpler and easier to follow again.

Each sub-graph right of "edit" generally has higher external validity than the previous at the cost of taking more active and passive time as well as having lower internal validity (more likely to be flaky).
This introduces another guiding principle; the mean-time-to-feedback, especially negative feedback, should be minimized.

`cargo check` or similar can replace `cargo test` when no host unit
tests exist. This can often be easily configured to run in an IDE where
failures to compile come with links into the code that are easy to
follow.

Removing the app in the validation step is important in order to reset
state such as `localdata` that could affect test results.

Building the app with the release profile is important for the
validation because some bugs present only with specific build
configurations.